+++ src/vue-form-state.vue
@@ -12,7 +12,7 @@
 </template>
 
 <script>
-    import useFormState from './vue-form-state.composite';
+    import useFormState from './use-form-state.composite';
 
     export default {
         props: {
